+++ Modules/quote/View/Quotelist/editnewuserAction.php
@@ -43,7 +43,7 @@ if ($id_quote > 0) {
     ?>
         <div class="col-12 pm-form">
             <br/>
-            <a class="btn btn-primary" id="additembutton"><?php echo QuoteTranslator::NewItem($lang) ?></a>
+            <a onclick="addItem()" class="btn btn-primary" id="additembutton"><?php echo QuoteTranslator::NewItem($lang) ?></a>
             <a class="btn btn-danger" href="quotepdf/<?php echo $id_space.'/'.$id_quote ?>"><?php echo QuoteTranslator::PDF($lang) ?></a>
 
             <?php echo $tableHtml ?>
@@ -52,18 +52,60 @@ if ($id_quote > 0) {
     }
     ?>
 
-    <!--  *************  -->
-    <!--  Popup windows  -->
-    <!--  *************  -->
-    <link rel="stylesheet" type="text/css" href="Framework/pm_popup.css">
-    <div id="hider" class="col-12"></div> 
-    <div id="itemspopup_box" class="pm_popup_box" style="display: none;">
-        <div class="col-1 offset-11" style="text-align: right;"><a id="itemsbuttonclose" class="bi-x-circle-fill" style="cursor:pointer;"></a></div>
-            <?php echo $formitemHtml ?>
-    </div> 
+
+    <div id="entriespopup_box" class="modal" tabindex="-1">
+    <div class="modal-dialog modal-lg">
+        <div class="modal-content">
+        <div class="modal-header">
+            <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+        </div>
+        <div class="modal-body">
+        <?php echo $formitemHtml ?>
+        </div>
+        </div>
+    </div>
+    </div>
 
 
-    <?php include 'Modules/quote/View/Quotelist/editnewscript.php'; ?>
+    <script>
+
+    function editentry(id) {
+            var arrayid = id.split("_");
+            showEditForm(arrayid[1]);        
+    }
+
+    function addItem() {
+        $('#id_quote').val(<?php echo $id_quote ?>);
+        $('#id').val(0);
+        $('#id_item').val(0);
+        $('#quantity').val("");
+        $('#comment').val("");
+
+        let myModal = new bootstrap.Modal(document.getElementById('entriespopup_box'))
+        myModal.show();
+    }
+
+    function showEditForm(id) {
+        $.post(
+                'quotegetitem/<?php echo $id_space ?>/' + id,
+                {},
+                function (data) {
+
+                    $('#id_quote').val(<?php echo $id_quote ?>);
+                    $('#id').val(data.id);
+                    $('#id_item').val(data.module + "_" + data.id_content);
+                    $('#quantity').val(data.quantity);
+                    $('#comment').val(data.comment);
+
+                    let myModal = new bootstrap.Modal(document.getElementById('entriespopup_box'))
+                    myModal.show();
+                },
+                'json'
+                );
+
+    }
+
+    </script> 
     
 <?php endblock(); ?>